fixed playerHealth getter in other player scripts

+++ Assets/PlayerHealth.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PlayerHealth : MonoBehaviour
 {
-    public double hp = 100;
+    public int hp = 100;
     // Start is called before the first frame update
     void Start()
     {
